Make some modifications in productmanagement service
Source : Admin.java (mod.), ProductManagementServlet.java (mod.), productmanagement.jsp (mod.), productmanagement.js (mod.)

Class : javaPackage.Admin (mod.), servletPackage.ProductManagementServlet (mod.)

Function :
- (Admin)setOrderLimit() : Calculate quantity limited per order of a product and obtain orderLimit as the  rounded-off value of what multiple of the base quantity is the limited quantity

Functionalities added :
- Store orderLimit of product while adding or editing product
- Change MaxQty from integer to decimal
- If MaxQty is whole number with only zeroes after decimal point, show only the integer part in edit product modal
- Modify image naming logic to avoid unnecessary String object creation and consecutive multiple underscores (_)
- During form validation through javascript, convert quantity values into integer before comparing

+++ WEB-INF/classes/ProductManagementServlet.java
@@ -65,7 +65,7 @@ public class ProductManagementServlet extends HttpServlet{
     int price = Integer.parseInt (priceS);
     String wght = multiReq.getParameter ("wght");
     String qty2S = multiReq.getParameter ("qty2");
-    int qty2 = Integer.parseInt (qty2S);
+    float qty2 = Float.parseFloat (qty2S);
     String unit2 = multiReq.getParameter ("unit2");
 
     String desc = multiReq.getParameter ("desc");
